Second commit

+++ Controller/PostsController.php
@@ -1,11 +1,10 @@
 <?php
 class PostsController extends AppController{
-	public $helpers = array('Html', 'Form', 'Session');
+	public $helpers = array('Html', 'Form', 'Session', 'Js');
 	public $components = array('Session');
 
-	public function index(){
-		$this->set('posts', $this->Post->find('all'));
-	}
+
+	
 
 	public function view($id = null){
 		if(!$id)
@@ -17,31 +16,45 @@ class PostsController extends AppController{
 		$this->set('post',$post);
 	}
 
-	public function add(){
-		if($this->request->is('post')){
-			$this->request->data['Post']['user_id'] = $this->Auth->user('id');
-			$this->Post->create();
-			if($this->Post->save($this->request->data)){
-				$this->Session->setFlash(__('Your post has been saved.'));
-				return $this->redirect(array('action' => 'index'));
-			}
-			$this->Session->setFlash(__('Unable to add your post'));
-		}
-	}
+	// public function add(){
+	// 	$thread_id = $this->Session->read('thread_id');
+	// 	if($this->request->is('post')){
+	// 		$user_id = $this->Auth->user('id');
+	// 		$this->loadModel('User');
+	// 		$user = $this->User->findById($user_id);
+	// 		$username = $user['User']['username'];
+
+	// 		$this->request->data['Post']['user_id'] = $user_id;
+	// 		$this->request->data['Post']['username'] = $username;
+	// 		$this->request->data['Post']['thread_id'] = $thread_id;
+
+	// 		$this->Post->create();
+	// 		if($this->Post->save($this->request->data)){
+	// 			$this->Session->setFlash(__('Your post has been saved.'));
+	// 			return $this->redirect(array('action' => 'index'));
+	// 		}
+	// 		$this->Session->setFlash(__('Unable to add your post'));
+	// 	}
+	// }
 
 	public function edit($id = null){
-		if(!$id)
+		if($id == null)
 			throw new NotFoundException(__('Invalid post'));
 
 		$post = $this->Post->findById($id);
+		$thread_id = $post['Post']['thread_id'];
+
 		if(!$post)
 			throw new NotFoundException(__('Invalid post'));
+
 		if($this->request->is(array('post', 'put'))){
+			
 			if($this->Post->save($this->request->data)){
 				$this->Session->setFlash(__('Your post has been updated'));
-				return $this->redirect(array('action' => 'index'));
+				return $this->redirect(array('controller'=>'threads','action' => 'listPost', $thread_id));
 			}
-			$this->Session->setFlash(__('Unable to update your post'));
+			else
+				$this->Session->setFlash(__('Unable to update your post'));
 		}
 		
 		if(!$this->request->data){
@@ -51,12 +64,19 @@ class PostsController extends AppController{
 	}
 
 	public function delete($id = null){
+		if($id == null)
+			throw new NotFoundException(__('Invalid post'));
+
 		if($this->request->is('get'))
 			throw new MethodNotAllowedException();
+
+		$post = $this->Post->findById($id);
+		$thread_id = $post['Post']['thread_id'];
+		//var_dump($thread_id);die();
+
 		if($this->Post->delete($id)){
-			$this->Session->setFlash(
-				__('The post has been deleted.'));
-			return $this->redirect(array('action' => 'index'));
+			$this->Session->setFlash(__('The post has been deleted.'));
+			return $this->redirect(array('controller'=>'threads','action' => 'listPost', $thread_id));
 		}
 			
 	}
@@ -65,7 +85,7 @@ class PostsController extends AppController{
 		//All registered users can add posts
 
 		if($this->action === 'index' || $this->action === 'add' || 
-			$this->action === 'view'){
+			$this->action === 'view' || $this->action === 'delete'){
 			return true;
 		}
 
@@ -81,16 +101,6 @@ class PostsController extends AppController{
 		return false;
 
 	}
-
-
-
-
-
-
-
-
-
-
